fixed progress bar in bulk upload - YES!!!! (when clearResults was called it was clearing it :( ) made is so chemistry attribute puts ppm and wt% under different headers (not sure if this is suppose to be like this) made it so chemical analysis amounts are double instead of float and made the corresponding changes to everything else made it so the sample cannot be found exception includes the alias if we have it made it so bulk upload does not show two errors for not finding a sample updated the analysis parser to sanitize the wt% and ppm fieldsu update the value in collection constraint to say what value was not found instead of all of the them made it so in bulk upload if we don't have a column or cell data those entries are not shown made it so that if no mineral is specified and the column does not equal "whole rock" we throw an error about missing the bulk rock property

+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/objects/list/ChemicalAnalysisListEx.java
@@ -169,7 +169,7 @@ public abstract class ChemicalAnalysisListEx extends ListEx<ChemicalAnalysis> {
 					ChemicalAnalysisProperty.total, true) {
 				protected Object getWidget(final MObject data,
 						final int currentRow) {
-					Float total = (Float) data.mGet(ChemicalAnalysisProperty.total);
+					Double total = (Double) data.mGet(ChemicalAnalysisProperty.total);
 					String text = "------";
 					if (total != null)
 						text = (ListExUtil.formatDouble(total,ListExUtil.defaultDigits));
