+++ mpdb-common/src/edu/rpi/metpetdb/client/model/SearchElement.java
@@ -4,8 +4,8 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class SearchElement implements IsSerializable {
 	private String elementSymbol;
-	private float lowerBound;
-	private float upperBound;
+	private Double lowerBound;
+	private Double upperBound;
 
 	public void setElementSymbol(String aSymbol) {
 		elementSymbol = aSymbol;
@@ -24,26 +24,26 @@ public class SearchElement implements IsSerializable {
 		return elementSymbol.toLowerCase();
 	}
 
-	public void setLowerBound(float aLowerBound) {
+	public void setLowerBound(Double aLowerBound) {
 		lowerBound = aLowerBound;
 	}
 
-	public float getLowerBound() {
+	public Double getLowerBound() {
 		return lowerBound;
 	}
 
-	public void setUpperBound(float aUpperBound) {
+	public void setUpperBound(Double aUpperBound) {
 		upperBound = aUpperBound;
 	}
 
-	public float getUpperBound() {
+	public Double getUpperBound() {
 		return upperBound;
 	}
 	
 	public void setValues(final String elementSymbol, final float lowerBound,
 			final float upperBound, final String measurementUnit){
 		this.elementSymbol = elementSymbol;
-		float unitMod = ChemicalAnalysis.getUnitOffset(measurementUnit);
+		Double unitMod = ChemicalAnalysis.getUnitOffset(measurementUnit);
 		this.lowerBound = lowerBound * unitMod;
 		this.upperBound = upperBound * unitMod;
 	}
